+++ tests/testthat/test-sensitivity-calculation.R
@@ -182,9 +182,28 @@ test_that("sensitivityCalculation fails with invalid `pkParameters`", {
       outputPaths = "Organism|PeripheralVenousBlood|Aciclovir|Plasma (Peripheral Venous Blood)",
       parameterPaths = parameterPaths
     ),
-    "Following non-standard PK parameters will not be calculated:\nabc\nxyz\n",
+    "Following PK parameters are specified but were not calculated:\nabc\nxyz\n",
     fixed = TRUE
   )
+})
+
+test_that("sensitivityCalculation works with user-defined `pkParameters`", {
+  # Create a new parameter based on the standard AUC parameter
+  myAUC <- addUserDefinedPKParameter(
+    name = "MyAUC",
+    standardPKParameter = StandardPKParameter$AUC_tEnd
+  )
+
+  results <- sensitivityCalculation(
+    simulation = simulation,
+    outputPaths = outputPaths,
+    parameterPaths = parameterPaths,
+    variationRange = variationRange,
+    pkParameters = c("C_max", "MyAUC")
+  )
+
+  expect_true(isOfType(results, "SensitivityCalculation"))
+  expect_equal(unique(results$pkData$PKParameter), c("C_max", "MyAUC"))
 })
 
 # Validate variationRange -------------------------------------------------